requirements.txt
Django
djangorestframework
djangorestframework-simplejwt
gunicorn
whitenoise
psycopg2-binary
python-dotenv
django-cors-headers
Pillow